+++ src/plugins/core/widgets/Message.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-const Message = ({ message = 'Write something witty' }) => {
+const Message = ({ message = 'Add something witty' }) => {
   return (
     <div className="Message">
       <h2>{message}</h2>